// src/components/layout/search/search-dialog/types.ts
/**
 * FilePath    : blog-client\src\components\layout\search\search-dialog\types.ts
 * Author      : jiaopengzi
 * Blog        : https://jiaopengzi.com
 * Copyright   : Copyright (c) 2025 by jiaopengzi, All Rights Reserved.
 * Description : 类型
 */

export interface SearchHistoryItem {
    id: number
    value: string
    time: number
}